test/e2e: Validate ReportDataSources and queries exist before running e2e test cases
We now group these e2e test cases using subtests to ensure common setup
code is run before running any individual test cases.
